+++ examples/c/select.c
@@ -17,7 +17,7 @@ int main(void) {
     uint64_t matching = 0;
 
     // Get the number of matching atoms from the frame
-    chfl_selection_evalutate(selection, frame, &matching);
+    chfl_selection_evaluate(selection, frame, &matching);
     printf("We have %"PRIu64" zinc in the frame\n", matching);
     chfl_match_t* matches = malloc((size_t)matching * sizeof(chfl_match_t));
 
@@ -32,7 +32,7 @@ int main(void) {
 
     // Create a selection for multiple atoms
     selection = chfl_selection("angles: name($1) H and name($2) O and name($3) H");
-    chfl_selection_evalutate(selection, frame, &matching);
+    chfl_selection_evaluate(selection, frame, &matching);
     printf("We have %"PRIu64" water in the frame\n", matching);
     matches = malloc((size_t)matching * sizeof(chfl_match_t));